// client/src/app/fluff/user.ts
export type Title = 'Dr.' | 'Prof.' | null;

// const <name>: <Typ> = <Inhalt>
const possibleTitles: Title[]=['Dr.', 'Prof.', null];

//An interface for user-data
export interface SimpleUser {
  title: Title;
  firstName: string;
  lastName: string
}

//An interface for user-data that includes an id
export interface IdentifiedUser extends SimpleUser {
  id: number
}

//An interface for user-data that includes a join-date
export interface User extends SimpleUser{
  readonly joinedAt: number //must be Unix Timestamp (generated with Date.now)
}